testtesttesttest

+++ cy_jobs/move_to_disk.py
@@ -23,6 +23,8 @@ from cy_xdoc.models.apps import App
 from cyx.cache_service.memcache_service import MemcacheServices
 cache = cy_kit.singleton(MemcacheServices)
 file_storage_path = config.file_storage_path
+if not hasattr(config,"tenants"):
+    raise Exception("tenants was not found in config, please call move_to_disk.py tenants='defaut,...'")
 if not os.path.isdir(file_storage_path):
     raise Exception(f"{file_storage_path} was not found")
 
@@ -62,8 +64,9 @@ def do_download_file(db: pymongo.database.Database, fs_id: bson.ObjectId, file_n
                     with open(full_file_path, "ab") as f:  # Replace with desired output filename
                         f.write(data)
 
-    except gridfs.errors.CorruptGridFile:
-        pass
+    except gridfs.errors.CorruptGridFile as ex:
+        print(f"{ex._message}")
+
     # fs.delete(fs_id)
 
 
@@ -78,31 +81,50 @@ def do_download_thumbs(db: pymongo.database.Database, available_thumbs, to_folde
             if fs_thumbs:
                 download_file_path = os.path.join(
                     to_folder,
-                    f"{file_name_only}_{x.split('/')[-1]}"
+                    f"{x.split('/')[-1]}"
 
                 )
-                with open(download_file_path, "wb") as f:  # Replace with desired output filename
-                    f.write(fs_thumbs.read())
+                if not os.path.isfile(download_file_path):
+                    with open(download_file_path, "wb") as f:  # Replace with desired output filename
+                        f.write(fs_thumbs.read())
+                else:
+                    print(f"{download_file_path} is existing")
                 # fs.delete(fs_thumbs._id)
             else:
                 print(f"{x} was not found in mongodb")
         except gridfs.errors.CorruptGridFile:
             pass
 
+from cyx.repository import Repository
+
+
+def is_bson_object(main_file_id):
+    import bson
+    if isinstance(main_file_id,bson.ObjectId):
+        return True
+    try:
+        bson.ObjectId(main_file_id)
+        return True
+    except:
+        return False
+
 
 def move_data(app_name: str):
+
     ret_count = 0
     file_context = mongodb_service.db(app_name).get_document_context(DocUploadRegister)
     agg = file_context.context.aggregate().sort(
         file_context.fields.RegisterOn.desc()
     ).match(
-        cy_docs.not_exists(file_context.fields.StoragePath) & (file_context.fields.Status == 1) & (file_context.fields.StorageType!="local")
+        (Repository.files.fields.IsExtracToDiskTime==None)
+        #cy_docs.not_exists(file_context.fields.StoragePath) & (file_context.fields.Status == 1) & (file_context.fields.StorageType!="local")
 
     ).limit(10)
     for x in agg:
         try:
 
             main_file_id = x[file_context.fields.MainFileId]
+
             main_thumb_id = x[file_context.fields.ThumbFileId]
             register_on: datetime.datetime = x[file_context.fields.RegisterOn]
             file_name = x[file_context.fields.FullFileNameLower]
@@ -135,33 +157,47 @@ def move_data(app_name: str):
             )
             if not os.path.isdir(full_path):
                 os.makedirs(full_path, exist_ok=True)
-            fs_id = main_file_id
-            if isinstance(fs_id, str):
-                try:
-                    fs_id = bson.ObjectId(fs_id)
-                except bson.errors.InvalidId:
-                    continue
-            do_download_file(
-                db=file_context.__client__.get_database(file_context.__db_name__),
-                fs_id=fs_id,
-                file_name=file_name,
-                to_folder=full_path
-            )
-            if main_thumb_id:
+            if is_bson_object(main_file_id):
+                fs_id = bson.ObjectId(main_file_id)
+            else:
+                fs_item= file_context.__client__.get_database(file_context.__db_name__).get_collection("fs.files").find_one(
+                    {"filename":{"$regex": f"^{upload_id}/"}}
+                )
+                if fs_item:
+                    fs_id = fs_item.get("_id")
+                    Repository.files.app(app_name).context.update(
+                        Repository.files.fields.id==upload_id,
+                        Repository.files.fields.MainFileId << str(fs_id)
+                    )
+
+            if not os.path.isfile(os.path.join(full_path,file_name)):
+                do_download_file(
+                    db=file_context.__client__.get_database(file_context.__db_name__),
+                    fs_id=fs_id,
+                    file_name=file_name,
+                    to_folder=full_path
+                )
+            else:
+                print(f'{os.path.join(full_path,file_name)} is existing')
+            if main_thumb_id and not os.path.isfile(os.path.join(full_path,f"{file_name}.webp")):
                 if isinstance(main_thumb_id, str):
                     try:
                         main_thumb_id = bson.ObjectId(main_thumb_id)
+                        print(main_thumb_id)
+                        do_download_file(
+                            db=file_context.__client__.get_database(file_context.__db_name__),
+                            fs_id=main_thumb_id,
+                            file_name=f"{file_name}.webp",
+                            to_folder=full_path
+                        )
                     except bson.errors.InvalidId:
-                        continue
-                print(main_thumb_id)
-                do_download_file(
-                    db=file_context.__client__.get_database(file_context.__db_name__),
-                    fs_id=main_thumb_id,
-                    file_name=f"{file_name}.webp",
-                    to_folder=full_path
-                )
+                        pass
+            else:
+                print(f'{os.path.join(full_path,f"{file_name}.webp")} is existing')
+
             available_thumbs = x[file_context.fields.AvailableThumbs]
             if isinstance(available_thumbs, list):
+
                 do_download_thumbs(db=file_context.__client__.get_database(file_context.__db_name__),
                                    available_thumbs=available_thumbs, to_folder=full_path, file_name=file_name)
 
@@ -173,7 +209,8 @@ def move_data(app_name: str):
             file_context.fields.id == upload_id,
             file_context.fields.StorageType << 'local',
             file_context.fields.StoragePath << f"local://{rel_path}/{file_name}",
-            file_context.fields.MainFileId << f"local://{rel_path}/{file_name}"
+            file_context.fields.IsExtracToDiskTime << "1"
+            # file_context.fields.MainFileId << f"local://{rel_path}/{file_name}"
         )
         cache.clear_all()
         ret_count += 1
@@ -204,9 +241,10 @@ apps = mongodb_service.db("admin").get_document_context(App).context.find(
 #     filter,linmit=10
 # )
 # mime_type:str = item[qr.fields.MimeType]
-
+tenants = config.tenants.split(',')
 while True:
-    for x in apps:
-        app_name = x[app_qr.fields.NameLower] or x[app_qr.fields.Name]
+    for x in tenants:
+
+        app_name = x
         do_move_all(app_name)
